+++ plugins/global.js
@@ -1,4 +1,4 @@
 import Vue from "vue";
-import PageLayout from "@/components/PageLayout.vue";
+import PageLayout from "@/components/layouts/PageLayout.vue";
 
 Vue.component("page-layout", PageLayout);
